Fix terminals for launch V3 FCs
